configure: fix shared library suffix on macOS
See <https://lists.gnu.org/archive/html/libtool/2010-04/msg00020.html>,
shrext_cmds is a command to be evaluated. Linux uses .so for both
loadable and shared libraries (in terms of libtool), so the problem is
not visible there, but on macOS this caused configure to literally look
for

libnspr4`test .$module = .yes && echo .so || echo .dylib`

for the nss backend, while nss provides libnspr4.dylib.