travis: enable clang sanitizers
These found useful things (see e.g. #220), but it's even better if these
are ran all the time to catch badness early.

Require an explicit newer clang version as some checks (e.g. local
bounds) are not available with the default 4 years old toolchain. (And
clang depends on gcc to some extent, so use a modern version of both.)